pyproject.toml
[tool.poetry]
name = "streamlit-tradier"
version = "0.1.0"
description = ""
authors = ["dixon0105 <dixonwan@u.nus.edu>"]
readme = "README.md"
packages = [{include = "streamlit_tradier"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.3.0"
streamlit-chat = "^0.0.2.2"
qrcode = "^7.4.2"
psycopg2-binary = "^2.9.6"
PyYAML = "^6.0"
streamlit-authenticator = "^0.2.1"
pyotp = "^2.8.0"
loguru = "^0.7.0"
pydantic = "^1.10.7"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pylint = "^2.17.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
